+++ src/index.ts
@@ -4,6 +4,7 @@
 import {CommonModule} from '@angular/common'
 import {NgModule} from '@angular/core'
 
+import {InputInjector} from './components/inputInjector/inputInjector.component'
 import {InputsModule} from './inputs/inputs.module'
 import validators from './validators'
 
@@ -12,11 +13,14 @@ import validators from './validators'
 		CommonModule,
 		InputsModule
 	],
-	exports: [InputsModule]
+	declarations: [
+		InputInjector
+	],
+	exports: [InputInjector, InputsModule]
 })
 export class RamsterUIFormsModule {}
 
-export {InputsModule, validators}
+export {InputInjector, InputsModule, validators}
 export * from './inputs/autocomplete/autocomplete.interfaces'
 export * from './inputs/base/baseInput.component'
 export * from './inputs/base/baseInput.interfaces'
@@ -27,3 +31,4 @@ export * from './inputs/input/input.interfaces'
 export * from './inputs/select/select.interfaces'
 export * from './inputs/slideToggle/slideToggle.interfaces'
 export * from './inputs/textarea/textarea.interfaces'
+export * from './services/formBuilder'